pyproject.toml
[tool.poetry]
name = "sailors-log"
version = "0.1.0"
description = ""
authors = ["RincewindWizzard <git@magierdinge.de>"]
license = "LGPL"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
gpxpy = "^1.6.2"
geopy = "^2.4.1"
django-background-tasks = "^1.2.8"
requests = "^2.32.4"
pytest = "^8.4.0"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pytest-django = "^4.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
